Add dump method to bufr
